% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ANN_new_sample}
\alias{ANN_new_sample}
\title{Artificial neural network (ANN) classification of a new sample}
\usage{
ANN_new_sample()
}
\arguments{
\item{fasta.file_path}{FASTA file path (e.g. FASTA file viral.genomic.fna downloaded from NCBI: ftp://ftp.ncbi.nih.gov/refseq/release/viral).}

\item{taxonomy.file_path}{Taxonomy file path (e.g. taxonomy file taxonomy_viruses_available.csv delivered within this package).}

\item{temp.directory}{Results directory containing accuracy results, generalised confusion matrix results and accuracy as well as loss graphics of the simulation runs.}

\item{read_sequences.file_path}{Only read sequences without identifier or species names, extracted from FASTQ or SAM file!}

\item{model.file_path}{File path of ANN model trained on artificially generated data.}

\item{predictions.file_path}{File path of the result file of predicted taxonomic orders.}
}
\description{
Step 2 of the taxa classification and estimation procedure.
This function loads an ANN model which was trained on artificial data,
computes input features of the new, adjusted sample file (FASTQ or SAM)
and stores the predicted classes of its read sequences.
}
\examples{
# Please specify your file paths and directories!
fasta.file_path = "~/ag_bioinf/genomes/viruses_na/refseq/viral.genomic.fna" # Download from NCBI!
taxonomy.file_path = "inst/extdata/taxonomy_viruses_available.csv" # Relative file path
temp.directory = "~/ag_bioinf/research/metagenomics/temp" # Results directory
read_sequences.file_path = "~/ag_bioinf/research/metagenomics/Data/Seehund_Mapping/read_sequences.txt"
model.file_path = "inst/extdata/model_training_1_dataset.h5"
predictions.file_path = "~/ag_bioinf/research/metagenomics/temp/900000_training_samples_1_iteration/test/predictions_seal_sample.rds"
\dontrun{
ANN_new_sample ( fasta.file_path, taxonomy.file_path, temp.directory, read_sequences.file_path )
}
}
\seealso{
\code{\link{class_specific_PPVs}}, \code{\link{estimation_taxa_distribution}}
}
