% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{estimation_taxa_distribution}
\alias{estimation_taxa_distribution}
\title{Prior and posterior estimation of taxa distribution}
\usage{
estimation_taxa_distribution()
}
\arguments{
\item{temp.directory}{Results directory containing accuracy results, generalised confusion matrix results and accuracy as well as loss graphics of the simulation runs.}

\item{graphics.directory}{Directory of graphics result file.}

\item{a_priori_table.file_path}{Predictions of previously classified reads of new sample file.}
}
\description{
Step 3 of the taxa classification and estimation procedure.
This function loads the predicted classes of the ANN model which was trained on artificial data,
computes prior as well as posterior taxa distribution estimations
and saves a graphics result file containing the estimation of the predicted classes.
}
\examples{
# Please specify your file paths and directories!
temp.directory = "inst/extdata/class_specific_PPVs_results/" # Results directory
graphics.directory = "~/ag_bioinf/research/metagenomics/ManuscriptNeuralNet/Graphics/"
a_priori_table.file_path = "inst/extdata/predictions_3154562.rds"
\dontrun{
estimation_taxa_distribution ( temp.directory, graphics.directory, a_priori_table.file_path )
}
}
\seealso{
\code{\link{class_specific_PPVs}}, \code{\link{ANN_new_sample}}
}
